requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
groq==0.13.0
numpy==2.2.3
python-dotenv==1.1.0
